length is defined as an unsigned two-byte number
The message data block can have any length from 0 to 65535 therefore an
uint16_t is required to pass the length correctly to the callback
function.

+++ src/Opc.h
@@ -53,6 +53,6 @@ struct OpcClient {
   uint8_t* buffer;
 };
 
-typedef void (*OpcMsgReceivedCallback)(uint8_t channel, uint8_t command, uint8_t length, uint8_t* data);
+typedef void (*OpcMsgReceivedCallback)(uint8_t channel, uint8_t command, uint16_t length, uint8_t* data);
 typedef void (*OpcClientConnectedCallback)(WiFiClient&);
 typedef void (*OpcClientDisconnectedCallback)(OpcClient&);
